+++ common/migrations/m180612_030312_create_notice.php
@@ -20,17 +20,17 @@ class m180612_030312_create_notice extends Migration
 
         $this->createTable('{{%notice}}', [
             'notice_id'      =>Schema::TYPE_INTEGER.'(11) UNSIGNED PRIMARY KEY AUTO_INCREMENT NOT NULL COMMENT "主键"',
-            'user_id'         => $this->integer(11)->unsigned()->notNull()->comment('用户id'),
+            'user_id'         => $this->integer()->unsigned()->notNull()->comment('用户id'),
             'title'           => $this->string(50)->notNull()->defaultValue(' ')->comment('标题'),
             'content'         => $this->text()->notNull()->comment('内容'),
             'type'            => $this->tinyInteger(2)->unsigned()->notNull()->defaultValue(1)->comment(' 类型：1:通知,2:消息,3:待办'),
             'related_route'  => $this->string(100)->notNull()->defaultValue(' ')->comment('跳转用的相关路由'),
-            'related_id'     => $this->integer(11)->unsigned()->notNull()->defaultValue(0)->comment('跳转到的对应ID'),
+            'related_id'     => $this->integer()->unsigned()->notNull()->defaultValue(0)->comment('跳转到的对应ID'),
             'tips'            => $this->string(20)->notNull()->defaultValue(' ')->comment(' 额外提示'),
             'tips_level'     => $this->tinyInteger(1)->unsigned()->notNull()->defaultValue(1)->comment(' 额外提示程度：1:todo,2:urgent,3:doing,4:processing'),
             'status'         => $this->tinyInteger(1)->unsigned()->notNull()->defaultValue(0)->comment('激活状态:10为启用，0位禁用'),
-            'created_at'    => $this->integer(11)->unsigned()->notNull()->defaultValue(0)->comment('创建时间戳'),
-            'updated_at'    => $this->integer(11)->unsigned()->notNull()->defaultValue(0)->comment('修改时间戳'),
+            'created_at'    => $this->integer()->unsigned()->notNull()->defaultValue(0)->comment('创建时间戳'),
+            'updated_at'    => $this->integer()->unsigned()->notNull()->defaultValue(0)->comment('修改时间戳'),
         ], $tableOptions);
     }
