+++ common/migrations/m180612_032913_create_config.php
@@ -23,8 +23,8 @@ class m180612_032913_create_config extends Migration
             'symbol'        => $this->string(20)->notNull()->comment('标识'),
             'content'       => $this->text()->notNull()->comment('值'),
             'encode'        => $this->tinyInteger(1)->unsigned()->notNull()->defaultValue(1)->comment('值的编码形式1:string,2:json,3:int'),
-            'created_at'    => $this->integer(11)->unsigned()->notNull()->defaultValue(0)->comment('创建时间戳'),
-            'updated_at'    => $this->integer(11)->unsigned()->notNull()->defaultValue(0)->comment('修改时间戳'),
+            'created_at'    => $this->integer()->unsigned()->notNull()->defaultValue(0)->comment('创建时间戳'),
+            'updated_at'    => $this->integer()->unsigned()->notNull()->defaultValue(0)->comment('修改时间戳'),
         ], $tableOptions);
     }
